requirements.txt
Django==4.0.6
djangorestframework==3.13.1
numpy==1.23.1
psycopg2==2.9.3
face_recognition==1.3.0
opencv-python==4.6.0.66
scikit-learn==1.1.1